refactor: move Module definition to module.hpp
